pyproject.toml
[project]
name = "talkter"
version = "0.1.0"
description = ""
authors = [
    {name = "ChanHyuk Park",email = "devch1013@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "ruff (>=0.14.1,<0.15.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "drf-yasg (>=1.21.11,<2.0.0)",
    "firebase-admin (>=7.1.0,<8.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "psycopg[binary] (>=3.2.11,<4.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "boto3 (>=1.40.55,<2.0.0)",
    "django-storages (>=1.14.6,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "beautifulsoup4 (>=4.14.2,<5.0.0)",
    "playwright (>=1.55.0,<2.0.0)",
    "pdf2image (>=1.17.0,<2.0.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "pymupdf (>=1.26.5,<2.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "langgraph (>=1.0.1,<2.0.0)",
    "langchain (>=1.0.1,<2.0.0)",
    "ipykernel (>=7.0.1,<8.0.0)",
    "langchain-anthropic (>=1.0.0,<2.0.0)",
    "langchain-openai (>=1.0.0,<2.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)"
]

[tool.poetry]
package-mode = false  # 이 줄 추가

[tool.poetry.group.dev.dependencies]
ipykernel = "^7.0.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.dev"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "-v -s --tb=short --reuse-db --nomigrations"